moved key and ignore fields from the controls reducer to the dataset reducer
