+++ src/domain/controls.test.js
@@ -1,17 +1,9 @@
-import { createEpicMiddleware } from 'redux-observable';
-import configureMockStore from 'redux-mock-store';
-import rootEpic from 'epics/root-epic'
 import { 
-  default as dataset,
-  setDataset 
+  default as dataset
 } from "./dataset";
 import {
   default as controls,
   setHierarchyConfig,
-  setKeyFields,
-  getKeyFields,
-  setIgnoredFields,
-  getIgnoredFields,
   showNodes,
   useDarkTheme,
   colorBy,
@@ -36,49 +28,8 @@ describe("Controls reducer", () => {
         hierarchyConfig: hierarchyConfig,
         shouldShowNodes: true,
         darkTheme: false,
-        colorBy: null,
-        keyFields: [],
-        ignoredFields: []
+        colorBy: null
       });
-
-      done();
-    });
-  });
-
-  describe("setKeyFields", () => {
-    it("sets the fields to use as a key for comparison", (done) => {
-      const keys = [{ path: ["uid"], displayName: "UID" }];
-      const action = setKeyFields(keys);
-      const owner = uuidv4();
-      const ds = [
-        { 'uid': "uid1", 'role': { 'role': "role", 'confidence': 80 } },
-        { 'uid': "uid2", 'role': { 'role': "role", 'confidence': 80 } }
-      ];
-      const configuration = {
-        fields: [
-          { 'path': ["uid"], 'displayName': "UID", 'groupable': true },
-          { 'path': ["role", "role"], 'displayName': "Role", 'groupable': false }
-        ]
-      };
-      const dataset = {
-        datasets:[]
-      }
-      dataset.datasets[owner] ={ 'dataset': ds, 'configuration': configuration }
-      const result = reducer( {dataset: dataset}, action);
-
-      expect(getKeyFields(result)).to.deep.equal(keys);
-
-      done();
-    });
-  });
-
-  describe("setIgnoreFields", () => {
-    it("sets the fields to ignore in comparison", (done) => {
-      const ignored = [{ path: ["timestamp"], displayName: "Timestamp" }];
-      const action = setIgnoredFields(ignored);
-      const result = reducer({}, action);
-
-      expect(getIgnoredFields(result)).to.deep.equal(ignored);
 
       done();
     });
