+++ src/features/visualization/d3-viz/setup-annotations.js
@@ -34,7 +34,7 @@ const mapNodesToAnnotationArray = (nodes, colorMap) =>{
     'color': colorMap[d.data.fieldValue] && !colorMap[d.data.fieldValue].disabled ? 
                 colorMap[d.data.fieldValue].color : "black"
   }));
-
+  console.log("annotation array: %o", annotations);
   return annotations;
 }
